backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-api-python-client==2.116.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
youtube-transcript-api==0.6.1
yt-dlp>=2024.3.10
celery==5.3.6
redis==5.0.1
pillow==10.2.0